Divide models into separate files
