+++ apps/courses/models/__init__.py
@@ -1,0 +1,11 @@
+# HTK Imports
+from htk.apps.courses.models.chapter import Chapter
+from htk.apps.courses.models.course import Course
+from htk.apps.courses.models.lesson import Lesson
+
+
+__all__ = [
+    'Chapter',
+    'Course',
+    'Lesson',
+]
