+++ apps/courses/models/assessment.py
@@ -1,0 +1,8 @@
+# HTK Imports
+from htk.models.classes import HtkBaseModel
+
+
+class Assessment(HtkBaseModel):
+    """Assessments are quizzes, exams, etc"""
+
+    pass
